+++ Domain/src/main/scala/peal/synthesis/Condition.scala
@@ -6,6 +6,6 @@ import peal.domain.PolicySet
 
 trait Condition {
   def getPol : PolicySet
-  def getTh : Double
+  def getTh : BigDecimal
   def synthesis(z3: Z3Context, consts: Map[String, Z3AST]): Z3AST
 }
